+++ internal/database/search_contexts_mock.go
@@ -10,5 +10,5 @@ type MockSearchContexts struct {
 	GetSearchContext                    func(ctx context.Context, opts GetSearchContextOptions) (*types.SearchContext, error)
 	ListSearchContextsByUserID          func(ctx context.Context, userID int32) ([]*types.SearchContext, error)
 	ListInstanceLevelSearchContexts     func(ctx context.Context) ([]*types.SearchContext, error)
-	GetSearchContextRepositoryRevisions func(ctx context.Context, searchContextID int32) ([]*types.SearchContextRepositoryRevisions, error)
+	GetSearchContextRepositoryRevisions func(ctx context.Context, searchContextID int64) ([]*types.SearchContextRepositoryRevisions, error)
 }
